refactor: separate codebase into two modules
